Refactor Uniswap V2/V3 to RPC Poller
